Add a new struct Transformation
This struct holds the matrix and handles all of the
transformation methods. This will be used on the Scene and on the
Geometry.
